Fix format strings.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1486 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ textblock.c
@@ -30,8 +30,10 @@ textblock_t *textblock_init(void)
     cur_mem += mem;
     tot_mem += mem;
     max_mem = max(max_mem, cur_mem);
-    if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p %3d alloc, cur: %d, max: %d, tot: %d\n", 
-			       __FILE__,__LINE__, t, t->head, mem, cur_mem, max_mem, tot_mem );
+    if (DEBUG_TEXT(2))
+	fprintf(dbgout, "%s:%d  %p %p %3lu alloc, cur: %lu, max: %lu, tot: %lu\n", __FILE__,__LINE__, t, t->head,
+	    (unsigned long)mem, (unsigned long)cur_mem,
+	    (unsigned long)max_mem, (unsigned long)tot_mem);
     return t;
 }
 
@@ -50,8 +52,8 @@ void textblock_add(textblock_t *t, const char *text, size_t size)
     cur_mem += mem;
     tot_mem += mem;
     max_mem = max(max_mem, cur_mem);
-    if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p %3d add, cur: %d, max: %d, tot: %d\n", 
-			       __FILE__,__LINE__, cur, cur->data, cur->size, cur_mem, max_mem, tot_mem );
+    if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p %3lu add, cur: %lu, max: %lu, tot: %lu\n", 
+			       __FILE__,__LINE__, cur, cur->data, (unsigned long)cur->size, (unsigned long)cur_mem, (unsigned long)max_mem, (unsigned long)tot_mem );
     cur = cur->next = (textdata_t *) xcalloc(1, sizeof(textdata_t));
     t->tail = cur;
 }
@@ -63,8 +65,8 @@ void textblock_free(textblock_t *t)
     for (cur = t->head; (nxt = cur->next) != NULL; cur = nxt) {
 	mem = cur->size + sizeof(*cur);
 	cur_mem -= mem;
-	if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p %3d free, cur: %d, max: %d, tot: %d\n", 
-				   __FILE__,__LINE__, cur, cur->data, cur->size, cur_mem, max_mem, tot_mem );
+	if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p %3lu free, cur: %lu, max: %lu, tot: %lu\n", 
+				   __FILE__,__LINE__, cur, cur->data, (unsigned long)cur->size, (unsigned long)cur_mem, (unsigned long)max_mem, (unsigned long)tot_mem);
 	xfree((void*)cur->data);
 	xfree((void*)cur);
     }
@@ -72,9 +74,9 @@ void textblock_free(textblock_t *t)
     mem = sizeof(*t->head);
     cur_mem -= mem;
 
-    if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p free, cur: %d, max: %d, tot: %d\n", 
-			       __FILE__,__LINE__, t, t->head, cur_mem, max_mem, tot_mem );
+    if (DEBUG_TEXT(2)) fprintf(dbgout, "%s:%d  %p %p free, cur: %lu, max: %lu, tot: %lu\n", 
+			       __FILE__,__LINE__, t, t->head, (unsigned long)cur_mem, (unsigned long)max_mem, (unsigned long)tot_mem );
     xfree(t);
     cur_mem -= sizeof(t->head) + sizeof(t);
-    if (DEBUG_TEXT(1)) fprintf(dbgout, "cur: %d, max: %d, tot: %d\n", cur_mem, max_mem, tot_mem );
+    if (DEBUG_TEXT(1)) fprintf(dbgout, "cur: %lu, max: %lu, tot: %lu\n", (unsigned long)cur_mem, (unsigned long)max_mem, (unsigned long)tot_mem );
 }
